Taskを独立させる
