+++ frontend/src/modules/apiClient/tasks/index.ts
@@ -1,0 +1,7 @@
+export interface Task {
+  id: number;
+  name: string;
+  description: string;
+  created_at: string;
+  updated_at: string;
+}
